Add eslint and modify code to support asyn/await.
