+++ package.json
@@ -13,7 +13,7 @@
   "author": "Shaun",
   "license": "ISC",
   "engines": {
-    "node": "6.9.4"
+    "node": "7.6.0"
   },
   "dependencies": {
     "bluebird": "3.4.7",
@@ -39,6 +39,7 @@
     "ava": "0.17.0",
     "chai": "3.5.0",
     "coveralls": "2.11.15",
+    "eslint": "3.18.0",
     "nyc": "10.1.2",
     "sinon": "1.17.7"
   }
