+++ src/server/authorization/spotify/index.js
@@ -1,25 +1,24 @@
 'use strict'
 
 const express = require('express')
-const co = require('bluebird').coroutine
 
 const Constants = require('src/constants')
-const Spotify = require('src/spotify/app')
+const Spotify = require('src/spotify')
 const Kv = require('src/kv')
 const Logger = require('src/util/logger')
 
 const debug = Logger.getDebug('authorization')
-const router = express.Router({ mergeParams: true })
+const router = express.Router({mergeParams: true})
 
 router.get('/', (req, res) => {
-  return co(function* () {
+  return (async function () {
 
     let accessToken = Spotify.client.getAccessToken()
 
     if (accessToken) {
       debug(`Found Spotify access token in memory.`)
     } else {
-      accessToken = yield Kv.getAsync(Constants.STORAGE_KEY.SPOTIFY.ACCESS_TOKEN)
+      accessToken = await Kv.getAsync(Constants.STORAGE_KEY.SPOTIFY.ACCESS_TOKEN)
       Spotify.client.setAccessToken(accessToken)
     }
 
@@ -38,16 +37,16 @@ router.get('/authorize', (req, res) => {
     'playlist-read-private', 'playlist-read-collaborative',
     'playlist-modify-public', 'playlist-modify-private'
   ]
-  const state  = Date.now().toString()
+  const state = Date.now().toString()
   const authoriseURL = Spotify.client.createAuthorizeURL(scopes, state)
 
   res.redirect(authoriseURL)
 })
 
 router.get('/callback', (req, res) => {
-  return co(function* () {
+  return (async function () {
 
-    yield Spotify.authorize(req.query.code)
+    await Spotify.authorize(req.query.code)
     return res.redirect('/api/auth')
   })()
     .catch(err => res.send(err))
