+++ src/server/slack/router.js
@@ -3,21 +3,20 @@
 const express = require('express')
 const _ = require('lodash')
 const request = require('request-promise')
-const co = require('bluebird').coroutine
 
 const Logger = require('src/util/logger')
 const Constants = require('src/constants')
 const Controller = require('src/server/slack/controller')
-const Spotify = require('src/spotify/app')
-const Serializer =  require('src/server/slack/serializer')
+const Spotify = require('src/spotify')
+const Serializer = require('src/server/slack/serializer')
 
 const debug = Logger.getDebug('controller-slack')
-const router = express.Router({ mergeParams: true })
+const router = express.Router({mergeParams: true})
 
 router.use(Controller.authorizeSlack)
 
 router.post('/command', (req, res) => {
-  return co(function* () {
+  return (async function () {
 
     const text = _.get(req, 'body.text')
 
@@ -41,16 +40,16 @@ router.post('/command', (req, res) => {
         return res.send('l')
 
       default:
-        return res.json({ text: `We don't support that yet.` })
+        return res.json({text: `We don't support that yet.`})
     }
   })()
 })
 
 router.post('/interactive', (req, res) => {
 
-  return co(function* () {
+  return (async function () {
 
-    res.send({ text: Constants.MESSAGING.WAIT_FOR_IT })
+    res.send({text: Constants.MESSAGING.WAIT_FOR_IT})
 
     const callbackId = req.body.callback_id
 
@@ -63,40 +62,38 @@ router.post('/interactive', (req, res) => {
         debug(`Action received: ${JSON.stringify(action)}`)
 
         if (action.name === Constants.SLACK.INTERACTIVE.SONG_SEARCH.ACTION.SKIP_SEARCH.NAME) {
-          return { delete_original: true }
+          return {delete_original: true}
         }
 
         const trackId = action.value
-        const track = yield Spotify.getTrackFromId(trackId)
+        const track = await Spotify.getTrackFromId(trackId)
         const trackName = track.name
         const artists = Spotify.getArtistsFromTrack(track)
 
-        const result = yield Spotify.addTrackToTargetPlaylist(trackId)
+        const result = await Spotify.addTrackToTargetPlaylist(trackId)
 
         if (result.status !== Constants.RESULT.STATUS.SUCCESS) {
-          return { text: Constants.MESSAGING.UH_OH }
+          return {text: Constants.MESSAGING.UH_OH}
         }
 
-        return { text: Constants.MESSAGING.SPOTIFY.getSONG_SUCCESSFULLY_ADDED(trackName, artists) }
+        return {text: Constants.MESSAGING.SPOTIFY.getSONG_SUCCESSFULLY_ADDED(trackName, artists)}
 
       default:
         Logger.error(`Unsupported callbackId received: ${callbackId}.`)
-        return { text: Constants.MESSAGING.UH_OH }
+        return {text: Constants.MESSAGING.UH_OH}
     }
   })()
-    .then(res => request({ uri: req.body.response_url, method: 'POST', json: res }))
+    .then(res => request({uri: req.body.response_url, method: 'POST', json: res}))
 })
 
-function _searchMusicFromSpotify(requestBody, query, res) {
-  return co(function* () {
+async function _searchMusicFromSpotify(requestBody, query, res) {
 
-    res.json({ text: Constants.MESSAGING.WAIT_FOR_IT })
+  res.json({text: Constants.MESSAGING.WAIT_FOR_IT})
 
-    const results = yield Spotify.search(query)
-    const serialized = Serializer.serializeSearchResults(query, results.items)
+  const results = await Spotify.search(query)
+  const serialized = Serializer.serializeSearchResults(query, results.items)
 
-    yield request({ uri: requestBody.response_url, method: 'POST', json: serialized })
-  })()
+  await request({uri: requestBody.response_url, method: 'POST', json: serialized})
 }
 
 function _displayHelp(req, query, res) {
